Moved the tests to specific packages and created empty test files (to be tested ltr)
