+++ src/test/java/seedu/address/logic/commands/cheatsheet/FilterCheatSheetByTagCommandTest.java
@@ -1,0 +1,4 @@
+package seedu.address.logic.commands.cheatsheet;
+
+public class FilterCheatSheetByTagCommandTest {
+}
