+++ src/test/java/seedu/address/logic/commands/cheatsheet/ViewCheatSheetCommandTest.java
@@ -1,0 +1,4 @@
+package seedu.address.logic.commands.cheatsheet;
+
+public class ViewCheatSheetCommandTest {
+}
