+++ src/test/java/seedu/address/logic/commands/cheatsheet/ViewSpecificCheatSheetTagCommandTest.java
@@ -1,0 +1,4 @@
+package seedu.address.logic.commands.cheatsheet;
+
+public class ViewSpecificCheatSheetTagCommandTest {
+}
