+++ src/test/java/seedu/address/logic/parser/cheatsheet/DeleteCheatSheetCommandParserTest.java
@@ -1,0 +1,4 @@
+package seedu.address.logic.parser.cheatsheet;
+
+public class DeleteCheatSheetCommandParserTest {
+}
