+++ src/test/java/seedu/address/logic/parser/cheatsheet/EditCheatSheetCommandParserTest.java
@@ -1,0 +1,4 @@
+package seedu.address.logic.parser.cheatsheet;
+
+public class EditCheatSheetCommandParserTest {
+}
